refactor: move common directory to base and update imports
